backend/requirement.txt
flask
flask_sqlalchemy
flask_cors
openpyxl
pymysql
python-dotenv
flask_migrate